+++ clikt-mordant/src/commonTest/kotlin/com/github/ajalt/clikt/core/ContextTest.kt
@@ -206,7 +206,7 @@ class ContextTest {
             init {
                 context {
                     exitProcess = { status = it }
-                    echoer = { _, _, _, _ -> }
+                    echoMessage = { _, _, _, _ -> }
                 }
             }
         }